requirements.txt
asgiref==3.3.4
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-extensions==3.1.5
django-filter==21.1
pydotplus==2.0.2
psycopg2==2.9.2
